backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
firebase-admin>=6.4.0
gunicorn>=21.2.0
whitenoise>=6.6.0
Pillow>=10.2.0
